# docs/swagger.yaml
openapi: 3.0.2
info:
  title: yabarbank
  version: 1.0.0
paths:
  /users:
    summary: Path used to manage the list of users.
    description: >-
      The REST endpoint/path used to list and create zero or more `user` entities.  This path contains a
      `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `user` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
        required: true
      tags:
        - users
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                'User validation failed: {param}: Path `{param}` is required.':
                  value:
                    error: 'User validation failed: {param}: Path `{param}` is required.'
          description: Input validation failed. Some parameter is missing or in incorrect format.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Username already exists:
                  value:
                    error: Username already exists
          description: Username already exists
      operationId: createuser
      summary: Create a user
      description: Creates a new instance of a `user`.
  /sessions:
    summary: Path used to manage the list of sessions.
    description: >-
      The REST endpoint/path used to list and create zero or more `session` entities.  This path
      contains a `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `session` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/session'
            examples:
              POST /sessions request body:
                value:
                  username: TEST
                  password: TEST3434
        required: true
      tags:
        - sessions
      responses:
        '201':
          content:
            application/json:
              examples:
                POST /sessions 201 response:
                  value:
                    token: 615f573c763ab0e1cbd6080c
          description: Successful response.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid credentials:
                  value:
                    error: Invalid credentials
          description: Invalid credentials
      operationId: createsession
      summary: Create a session
      description: Creates a new instance of a `session`.
    delete:
      tags:
        - sessions
      responses:
        '204':
          description: Successful logout
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid Authorization header format:
                  value:
                    error: Invalid Authorization header format
                Missing Authorization header:
                  value:
                    error: Missing Authorization header
                Invalid token:
                  value:
                    error: Invalid token
          description: Invalid Authorization header
      security:
        -
          Bearer: []
      summary: Logs the user out
      description: Destroy the session
  /users/current:
    get:
      tags:
        - users
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
              examples:
                User details:
                  value:
                    accounts:
                      -
                        balance: 10000
                        createdAt: '2021-09-28T15:32:43.241Z'
                        currency: USD
                        id: 61533638c51acde23b70284f
                        name: Main
                        number: abcc06c2f7032eece45cfd0331afab3833d
                        userId: 61533638c51acde23b70284d
                    id: 61533638c51acde23b70284d
                    name: TEST
                    username: TEST
          description: |
            Returns user's details
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid Authorization header format:
                  value:
                    error: Invalid Authorization header
                Missing Authorization header:
                  value:
                    error: Missing Authorization header
          description: Problem with authorization header
      security:
        -
          Bearer: []
      summary: Get user's details
      description: Returns instance of `user`
  /transactions:
    summary: Path used to manage the list of transactions.
    description: >-
      The REST endpoint/path used to list and create zero or more `transaction` entities.  This path
      contains a `GET` and `POST` operation to perform the list and create tasks, respectively.
    get:
      tags:
        - transactions
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/transaction'
              examples:
                ' GET /transactions 200 response':
                  value:
                    accountFrom: 978c06c2f7032eece45cfd0331afab3833d
                    accountTo: 92c84bb50b65304f534480390cc90aafcbc
                    amount: 100
                    createdAt: '2021-10-21T18:38:02.678Z'
                    currency: EUR
                    receiverName: Kristi Maask
                    senderName: Daniil
                    status: Completed
          description: 'List of all user''s transactions with accounts from, names from, sums and date.'
        '401':
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/error'
                examples:
                  Invalid signature:
                    value:
                      error: Invalid token
            description: Invalid token
      security:
        -
          Bearer: []
      operationId: gettransactions
      summary: Returns list of transactions
      description: 'Returns all transactions of user with accounts from, names from, sums and date.'
    post:
      requestBody:
        description: A new `transaction` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/transaction'
        required: true
      tags:
        - transactions
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid amount:
                  value:
                    error: Invalid amount
                Missind parameter:
                  value:
                    error: 'Transaction validation failed: .* : Path `.*` is required.'
                'Path `accounFrom`  (`a`) is shorter than the minimum allowed length. ':
                  value:
                    error: 'Path `.*`  (`.*`) is shorter than the minimum allowed length. '
                'Transaction validation failed: explanation: Path `explanation` is required.':
                  value:
                    error: >-
                      Transaction validation failed: explanation: Path `explanation` is
                      required.
          description: Bad request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid Authorization header format:
                  value:
                    error: Invalid Authorization header
                Missing Authorization header:
                  value:
                    error: Missing Authorization header
          description: Problem with authorization header
        '402':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Insufficient funds:
                  value:
                    error: Insufficient funds
          description: Insufficient funds
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                accountFrom not found:
                  value:
                    error: accountFrom not found
          description: Not found
      security:
        -
          Bearer: []
      operationId: createtransaction
      summary: Creates transaction
      description: Creates a new instance of a transaction
  /transactions/b2b:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/jwt'
        required: true
      tags:
        - transactions
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/transaction'
              examples:
                POST /transactions/b2b response:
                  value:
                    receiverName: Julia Taro
          description: Transaction accepted
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid signature:
                  value:
                    error: Invalid signature
          description: Invalid signature
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Bank \w+ was not found in Central Bank:
                  value:
                    error: Bank \w+ was not found in Central Bank
          description: Bank was not found in Central Bank
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                There was an error communication with Central Bank:
                  value:
                    error: There was an error communication with Central Bank.*
          description: There was an error communication with Central Bank
      summary: Accepts transactions from other banks
  /transactions/jwks:
    description: ''
    get:
      tags:
        - transactions
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jwks'
          description: >-
            Set of keys in jwks format containing the public key used to verify JSON Web Token
            sent by the origin bank when initiating a transaction with another bank and signed
            with the corresponding private key
      summary: Returns JWKS for validating jwt-signatures
      description: >-
        Returns set of keys containing the public key used to verify JSON Web Token sent by the origin
        bank when initiating a transaction with another bank and signed with the corresponding private
        key
components:
  schemas:
    user:
      title: Root Type for user
      description: ''
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
      example:
        username: TEST
        password: TEST3434
        name: TEST
    error:
      title: Root Type for error
      description: ''
      type: object
      properties:
        error:
          type: string
      example:
        error: Error message.
    session:
      description: ''
      type: object
    transaction:
      title: Root Type for transaction
      description: ''
      type: object
      properties:
        accountFrom:
          type: string
        accountTo:
          type: string
        amount:
          format: int32
          type: integer
        explanation:
          type: string
      example:
        accountFrom: babc06c2f7032eece45cfd0331afab3833d
        accountTo: 093BY000000001
        amount: 1500000000
        explanation: Don’t go spend it all at once
    jwt:
      title: Root Type for jwt
      description: ''
      type: object
      properties:
        jwt:
          type: string
      example:
        jwt: >-
          eyJhbGciOiJSUzI1NiIsImtpZCI6IlFWaGxic05aWlk4M1lSRkhTLU9XSnFpcUVNYmpvYXFoWklpUFZSWmdXLWsifQ.eyJhY2NvdW50RnJvbSI6ImJhYmMwNmMyZjcwMzJlZWNlNDVjZmQwMzMxYWZhYjM4MzNkIiwiYWNjb3VudFRvIjoiYmFiMDcwZGRhYzhhYTk0NjZjNTVkOWI0NTEzZTlhNzg3NzQiLCJhbW91bnQiOjEwLCJjdXJyZW5jeSI6IlVTRCIsImV4cGxhbmF0aW9uIjoieHh4Iiwic2VuZGVyTmFtZSI6IlRFU1QifQ.B0UKNg9T9XPO-e7DiuyvT089AwqxhI3lQ8IsmLDtLHxzqHcbr3jAPq3eifpBfEmFU85s-ksTfZ8Pf-BmmYftRcmfcFjGL8aLUrcxf58fuNFySqjFOlIogVysFyeblqagOcJ7QJmT37rBYRTZ7pcuvQxlisdfDE8tS3aOXFD-xi8
    jwks:
      title: Root Type for jwks
      description: ''
      type: object
      properties:
        keys:
          type: array
          items:
            type: object
            properties:
              kty:
                type: string
              kid:
                type: string
              'n':
                type: string
              e:
                type: string
      example:
        keys:
          -
            kty: RSA
            kid: QVhlbsNZZY83YRFHS-OWJqiqEMbjoaqhZIiPVRZgW-k
            'n': >-
              RlWISHXpjWXdYGLSB99okpIW8eEL9ochaEMUyZ8bHW46CLXX-kkI8l6ue7KtxuyrWpuOpuzrHtZWp29H6fgyOq1mKZ0heK-XkV1rJoKHd-6wc0Jdpg2SPwC6BWRGcE2HnvHxT0s8pfa_GwTd7Tu27gIR2S7TVx7N86E5jxll7t0
            e: AQAB
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
tags:
  -
    name: sessions
    description: ''
  -
    name: users
    description: ''
  -
    name: transactions
    description: ''